[ travis ] using travis to build the project
We need a rather recent version of Agda so I fetch the one I build using
a travis instance running on the dev version which I host on dropbox.

The lipics cls needs to be patched to remove inputenc which is incompatible
with xelatex and would make the build fail otherwise.

Also, we need a patched version of polytable playing nice with unicode-
math (see: https://github.com/kosmikus/lhs2tex/issues/29).

Finally, index.md is used to generate a wee webpage pointing at the pdfs
and the html documentation.